Desarrollo cobros e impresion

+++ code/ajax/conf_inicial.php
@@ -1,12 +1,16 @@
 <?php
-
 	header('Content-Type: text/html; charset=utf-8');
 
-	$main_path = getenv('PATH_STORAGE') ?: '../..';
+	$main_path = getenv('PATH_STORAGE') ?: null;
 
-	if(file_exists("{$main_path}/conexion_inicial.txt")){
 //1. Elimina el archivo /conexion_inicial.txt si existe
-		unlink("{$main_path}/conexion_inicial.txt");
+	if( $main_path ){
+		if(file_exists("{$main_path}/conexion_inicial.txt")){
+			unlink("{$main_path}/conexion_inicial.txt");
+		}
+	}
+	if(file_exists("../../conexion_inicial.txt")){
+			unlink("../../conexion_inicial.txt");
 	}
 //2. Crea la cadena con los datos del formulario
 	$cadena_datos='';
@@ -43,17 +47,32 @@
 	$cadena_datos.= '<>'.$_POST['system_type'];
 				
 //3. Crea archivo /conexion_inicial.txt
-	$fp = fopen("{$main_path}/conexion_inicial.txt", "w");
-		fputs($fp,$cadena_datos);
-		fclose($fp);
+	if( $main_path ){
+		$fp = fopen("{$main_path}/conexion_inicial.txt", "w");
+			fputs($fp,$cadena_datos);
+			fclose($fp);
+	}
 
-	if(file_exists("{$main_path}/config.inc.php")){
+	$fp = fopen("../../conexion_inicial.txt", "w");
+	fputs($fp,$cadena_datos);
+	fclose($fp);
+
 //4. Elimina archivo /config.inc.php
-		unlink("{$main_path}/config.inc.php");
+	if( $main_path ){
+		if(file_exists("{$main_path}/config.inc.php")){
+			unlink("{$main_path}/config.inc.php");
+		}
+	}
+	if(file_exists("../../config.inc.php")){
+		unlink("../../config.inc.php");
 	}
 
 //5. Crea archivo /config.inc.php
-	$ini=fopen("{$main_path}/config.inc.php", "w");
+	
+	if( $main_path ){
+		$ini = fopen("{$main_path}/config.inc.php", "w");
+	}
+	$ini2 = fopen("../../config.inc.php", "w");
 	$datos="<?php\n";
 	$datos.="session_start();\n";
 	$datos.="//Definiciones de base de datos\n";
@@ -77,16 +96,31 @@
 	$datos.="	\$dur_session=0;//50000 modificado el Oscar 11.06.2018 para cerrar sesión al cerrar el explorador\n";
 	$datos.="	date_default_timezone_set('America/Mexico_City');\nheader('Content-Type: text/html; charset=utf-8');\n?>";
 	
-	fwrite($ini, $datos);
-	fclose($ini);
+	if( $main_path ){
+		fwrite($ini, $datos);
+		fclose($ini);
+	}
+
+	fwrite($ini2, $datos);
+	fclose($ini2);
 
 	
 //6. Elimina archivo /conexionDoble.php
-	if(file_exists("{$main_path}/conexionDoble.php")){
-		unlink("{$main_path}/conexionDoble.php");
+	if( $main_path ){
+		if(file_exists("{$main_path}/conexionDoble.php")){
+			unlink("{$main_path}/conexionDoble.php");
+		}
+	}
+
+	if(file_exists("../../conexionDoble.php")){
+		unlink("../../conexionDoble.php");
 	}
 //7. Crea archivo /conexionDoble.php
-	$ini=fopen("{$main_path}/conexionDoble.php", "w");
+	if( $main_path ){
+		$ini=fopen("{$main_path}/conexionDoble.php", "w");
+	}
+
+	$ini2=fopen("../../conexionDoble.php", "w");
 
 	$datos='';
 	$datos.="<?php\n";
@@ -153,9 +187,14 @@
 //fin de cambio Oscar 2023
 
 	$datos.="?>";
+	
+	if( $main_path ){
+		fwrite($ini, $datos);
+		fclose($ini);
+	}
 
-	fwrite($ini, $datos);
-	fclose($ini);
+	fwrite($ini2, $datos);
+	fclose($ini2);
 
 	echo 'ok';
 ?>
